Change git ignore/ignored/unignore to temporary ignoring.
The theory here is that there's too much extra info needed when adding
entries to the .gitignore files. For one, you need to determine if it
should be added to the project's .gitignore file, or the global ignore
file (typically ~/.gitignore). Second, you often want to add a wildcard
to the .gitignore file. This may not work as expected when run as a
command, because the shell will expand the wildcards before sending the
arguments to the command.