Vim: add Vundle to manage plugins.
Chose Vundle because it allows us to just declare the names of packages
in the Vim config files (and run `:PluginInstall`). This makes it much
easier to deal with plugins -- we can just add them to the config file,
commit to git, and run `:PluginInstall!`. We don't have to update our
installation scripts, just our config files.