Ripgrep (rg): Pretty output; default piping through `less`
Ripgrep doesn't have a config file, so we have to use a shell alias (actually
a shell function in this case) to set any defaults. We have to use a shell
function because the arguments need to come before the pipe. The `-t 1` test is
TRUE if STDOUT is an interactive terminal.