Add support for trusted ./bin to bundler auto-exec script
Note that I added `rails` and `spring` to the list, as they should be run from
`./bin` normally, or the global version if that doesn't exist.

The "trusted" bin directory idea comes from Tim Pope. To "trust" a directory,
just create a `safe` directory in the `.git` directory of your project. I use
a git alias (`git trust`) to do this.