Git: Add `pwip` (push work in progress) alias
This alias will push everything (including untracked files) in your working
directory to a separate branch on your origin (default) remote. The branch will
have a name indicating that it's a "WIP", with your username and the date.
The idea is to run this at the end of the day to ensure that you don't lose
any work. You can (and should) delete the branch from the remote the next day.

There's an additional `pwipf` alias, which is the same thing, but will force
push the branch, if you want to push a second time the same day.

Note that due to the way this works, if you have anything staged it will be
unstaged when the command completes.