Zsh prompt: Handle systems without `tput`
Assume standard ANSI escape sequences, which is pretty much the case for all
modern terminals in use.

OpenWRT does not have `tput`, and no package with it included. It should be in
a `ncurses-bin` package, but that package doesn't exist. Compiling it could be
an option, but I'm not ready to take on the challenge of installing a full
cross-compiler suite yet.